Feat team관련 jwt를 이용한 기본적인 권한 설정 추가

+++ src/main/java/com/iso/logus/domain/team/controller/TeamAuthController.java
@@ -1,7 +1,7 @@
 package com.iso.logus.domain.team.controller;
-
 import java.util.List;
 
+import javax.servlet.http.HttpServletRequest;
 import javax.validation.Valid;
 
 import org.springframework.http.HttpStatus;
@@ -15,8 +15,12 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.iso.logus.domain.team.domain.teamauth.AuthType;
 import com.iso.logus.domain.team.dto.TeamAuthDto;
 import com.iso.logus.domain.team.service.TeamAuthService;
+import com.iso.logus.domain.team.service.TeamUserService;
+import com.iso.logus.global.exception.AccessDeniedException;
+import com.iso.logus.global.jwt.JwtTokenProvider;
 
 import io.swagger.annotations.ApiImplicitParam;
 import io.swagger.annotations.ApiImplicitParams;
@@ -28,6 +32,8 @@ import lombok.RequiredArgsConstructor;
 public class TeamAuthController {
 	
 	private final TeamAuthService teamAuthService;
+	private final TeamUserService teamUserService;
+	private final JwtTokenProvider jwtTokenProvider;
 
 	@GetMapping(value = "/{teamId}")
 	@ApiImplicitParams({
@@ -42,7 +48,11 @@ public class TeamAuthController {
 		@ApiImplicitParam(name = "X-AUTH-TOKEN", value = "로그인 성공 후 access_token", required = false, dataType = "String", paramType = "header")
 	})
 	@ResponseStatus(code = HttpStatus.CREATED)
-	public void createTeamAuth(@Valid @RequestBody TeamAuthDto.SaveRequest saveRequest) {
+	public void createTeamAuth(@Valid @RequestBody TeamAuthDto.SaveRequest saveRequest, HttpServletRequest request) {
+		String token = jwtTokenProvider.resolveToken(request);
+		String requestUid = jwtTokenProvider.getUserPk(token);
+		if(!teamUserService.isUserHasAuth(saveRequest.getTeamId(), requestUid, AuthType.authManageAuth))
+			throw new AccessDeniedException();
 		teamAuthService.createTeamAuth(saveRequest);
 	}
 	
@@ -50,7 +60,11 @@ public class TeamAuthController {
 	@ApiImplicitParams({
 		@ApiImplicitParam(name = "X-AUTH-TOKEN", value = "로그인 성공 후 access_token", required = false, dataType = "String", paramType = "header")
 	})
-	public void changeTeamAuth(@Valid @RequestBody TeamAuthDto.UpdateRequest updateRequest) {
+	public void changeTeamAuth(@Valid @RequestBody TeamAuthDto.UpdateRequest updateRequest, HttpServletRequest request) {
+		String token = jwtTokenProvider.resolveToken(request);
+		String requestUid = jwtTokenProvider.getUserPk(token);
+		if(!teamUserService.isUserHasAuth(updateRequest.getTeamId(), requestUid, AuthType.authManageAuth))
+			throw new AccessDeniedException();
 		teamAuthService.changeTeamAuth(updateRequest);
 	}
 	
@@ -58,7 +72,11 @@ public class TeamAuthController {
 	@ApiImplicitParams({
 		@ApiImplicitParam(name = "X-AUTH-TOKEN", value = "로그인 성공 후 access_token", required = false, dataType = "String", paramType = "header")
 	})
-	public void deleteTeamAuth(@PathVariable long teamId, @PathVariable String authName) {
+	public void deleteTeamAuth(@PathVariable long teamId, @PathVariable String authName, HttpServletRequest request) {
+		String token = jwtTokenProvider.resolveToken(request);
+		String requestUid = jwtTokenProvider.getUserPk(token);
+		if(!teamUserService.isUserHasAuth(teamId, requestUid, AuthType.authManageAuth))
+			throw new AccessDeniedException();
 		teamAuthService.deleteTeamAuth(teamId, authName);
 	}
 }
