+++ src/main/java/com/iso/logus/domain/team/controller/TeamController.java
@@ -2,6 +2,7 @@ package com.iso.logus.domain.team.controller;
 
 import java.util.List;
 
+import javax.servlet.http.HttpServletRequest;
 import javax.validation.Valid;
 
 import org.springframework.http.HttpStatus;
@@ -15,9 +16,13 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.iso.logus.domain.team.domain.teamauth.AuthType;
 import com.iso.logus.domain.team.dto.TeamDto;
 import com.iso.logus.domain.team.service.TeamSearchService;
 import com.iso.logus.domain.team.service.TeamService;
+import com.iso.logus.domain.team.service.TeamUserService;
+import com.iso.logus.global.exception.AccessDeniedException;
+import com.iso.logus.global.jwt.JwtTokenProvider;
 
 import io.swagger.annotations.ApiImplicitParam;
 import io.swagger.annotations.ApiImplicitParams;
@@ -29,7 +34,9 @@ import lombok.RequiredArgsConstructor;
 public class TeamController {
 
 	private final TeamService teamService;
+	private final TeamUserService teamUserService;
 	private final TeamSearchService teamSearchService;
+	private final JwtTokenProvider jwtTokenProvider;
 	
 	@GetMapping
 	@ApiImplicitParams({
@@ -60,7 +67,11 @@ public class TeamController {
 	@ApiImplicitParams({
 		@ApiImplicitParam(name = "X-AUTH-TOKEN", value = "로그인 성공 후 access_token", required = false, dataType = "String", paramType = "header")
 	})
-	public void updateTeam(@PathVariable long id, @Valid @RequestBody TeamDto.UpdateRequest UudateRequest) {
+	public void updateTeam(@PathVariable long id, @Valid @RequestBody TeamDto.UpdateRequest UudateRequest, HttpServletRequest request) {
+		String token = jwtTokenProvider.resolveToken(request);
+		String requestUid = jwtTokenProvider.getUserPk(token);
+		if(!teamUserService.isUserHasAuth(id, requestUid, AuthType.teamNameAuth))
+			throw new AccessDeniedException();
 		teamService.updateTeam(id, UudateRequest);
 	}
 	
@@ -68,7 +79,11 @@ public class TeamController {
 	@ApiImplicitParams({
 		@ApiImplicitParam(name = "X-AUTH-TOKEN", value = "로그인 성공 후 access_token", required = false, dataType = "String", paramType = "header")
 	})
-	public void deleteTeam(@PathVariable long id) {
+	public void deleteTeam(@PathVariable long id, HttpServletRequest request) {
+		String token = jwtTokenProvider.resolveToken(request);
+		String requestUid = jwtTokenProvider.getUserPk(token);
+		if(!teamUserService.isUserHasMasterAuth(id, requestUid))
+			throw new AccessDeniedException();
 		teamService.deleteTeam(id);
 	}
 }
